# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/suppliers": {
            "get": {
                "description": "Retrieves a full list of registered suppliers from MongoDB",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "suppliers"
                ],
                "summary": "Get all supply chain vendors",
                "responses": {
                    "200": {
                        "description": "Returns wrapped supplier list array",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "array",
                                "items": {
                                    "$ref": "#/definitions/domain.Supplier"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal database query failure",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Validates and stores a new supplier profile into the system",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "suppliers"
                ],
                "summary": "Register a new vendor",
                "parameters": [
                    {
                        "description": "Supplier Profiles Payload",
                        "name": "supplier",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.Supplier"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Returns success message and supplier data",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Invalid payload error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "422": {
                        "description": "Business validation logic failure",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "domain.Supplier": {
            "type": "object",
            "properties": {
                "contact": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "is_active": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "tier": {
                    "description": "e.g., Tier-1 Premium, Tier-2 Standard",
                    "type": "string"
                }
            }
        }
    }
}